refactor: remove template_generation from planning module — proper standalone module
No backward-compat shim. All imports updated:
- planning/domain/generator.rs: deleted (was re-export shim)
- planning/domain/mod.rs: removed 'pub mod generator' + re-exports
- All planning files import from crate::template_generation::domain::X now
- 7 files updated: dto, factory, pipeline_factory_impl, pipeline_impl,
  service, repository, tests

The template_generation module is now the sole owner of:
  TemplateGenerator, ClaudeTemplateGenerator, GeneratorError, RepoContext,
  GeneratedTemplate, GeneratedTemplateCost, InvalidSymbolReference,
  ClaudeGeneratorConfig

889 tests pass.
